# docs/swagger.yaml
definitions:
  main.Cart:
    properties:
      created_at:
        type: string
      id:
        type: integer
      movie:
        description: relation
        type: string
      price:
        type: integer
      room:
        type: string
      seat:
        type: string
      seat_id:
        type: integer
      showtime_end:
        type: string
      showtime_id:
        type: integer
      showtime_start:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  main.CartFilter:
    properties:
      ids:
        items:
          type: integer
        type: array
      seat_ids:
        items:
          type: integer
        type: array
      showtime_ids:
        items:
          type: integer
        type: array
      user_ids:
        items:
          type: integer
        type: array
    type: object
  main.CartInput:
    properties:
      seat_id:
        type: integer
      showtime_id:
        type: integer
      user_id:
        type: integer
    type: object
...
  main.Paginate-main_Cart:
    properties:
      current_page:
        type: integer
      items:
        items:
          $ref: '#/definitions/main.Cart'
        type: array
      page_size:
        type: integer
      total_items:
        type: integer
      total_page:
        type: integer
    type: object
...
  main.Response-any:
    properties:
      data: {}
      message:
        type: string
    type: object
...
  main.Response-main_Cart:
    properties:
      data:
        $ref: '#/definitions/main.Cart'
      message:
        type: string
    type: object
...
  main.Response-main_Paginate-main_Cart:
    properties:
      data:
        $ref: '#/definitions/main.Paginate-main_Cart'
      message:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/carts:
    post:
      consumes:
      - application/json
      description: user create cart showtime
      parameters:
      - description: bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: body request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/main.CartInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Response-main_Cart'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.Response-any'
      summary: Create Cart
      tags:
      - carts
  /api/carts/{id}:
    delete:
      consumes:
      - application/json
      description: user delete cart by cart by id
      parameters:
      - description: bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: cart id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Response-any'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.Response-any'
      summary: Delete Cart
      tags:
      - carts
    get:
      consumes:
      - application/json
      description: user get cart by cart id
      parameters:
      - description: bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: cart id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Response-main_Cart'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.Response-any'
      summary: Get Cart
      tags:
      - carts
    put:
      consumes:
      - application/json
      description: user update cart by id
      parameters:
      - description: bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: cart id
        in: path
        name: id
        required: true
        type: integer
      - description: body request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/main.CartInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Response-main_Cart'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.Response-any'
      summary: Update Cart
      tags:
      - carts
  /api/carts/filter:
    post:
      consumes:
      - application/json
      description: user filter own carts
      parameters:
      - description: bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: pagination page
        in: query
        name: page
        type: integer
      - description: pagination page size
        in: query
        name: per_page
        type: integer
      - description: filter
        in: body
        name: request
        schema:
          $ref: '#/definitions/main.CartFilter'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Response-main_Paginate-main_Cart'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.Response-any'
      summary: Filter Cart
      tags:
      - carts
...
swagger: "2.0"